Update README.md
our voxel grid. Implement this carving process in carve(). Submit your code and a picture of what it looks like after one iteration of the carving. [15 points]

The last step in the pipeline is to carve out multiple views. Submit the nal output after all carvings have been completed, using num voxels 6; 000; 000. [5 points]

Notice that the reconstruction is not really that exceptional. This is because a lot of space is wasted when we set the initial bounds of where we carve. Currently, we initialize the bounds of the voxel grid to be the locations of the cameras. However, we can do better than this by completing a quick carve on a much lower resolution voxel grid (we use num voxels = 4000) to estimate how big the object is and retrieve tighter bounds. Complete the method get voxel bounds() and change the variable estimate better bounds in the main() func-tion to True. Submit your new carving, which should be more detailed, and your code. [10 points]

Finally, let’s have a fun experiment. Notice that in the rst three steps, we used perfect silhouettes to carve our object. Look at the estimate silhouette() function implemented and its output. Notice that this simple method does not produce a really accurate silhouette. However, when we run space carving on it, the result still looks decent!

Why is this the case? [2 points]

What happens if you reduce the number of views? [1 points]

What if the estimated silhouettes weren’t conservative, meaning that one or a few views had parts of the object missing? [2 points]

Representation Learning (10 points)